Выносит в отдельную функцию валидацию полей формы и функию нажатия Esc
