+++ js/form.js
@@ -1,9 +1,6 @@
-import { isEscEvent } from './utils.js';
-import { createHashTags, matchingPattern, isDuplicate, isTooMushHashTags } from './test-hash-tags.js';
-import { isValidityLengthComments } from './test-comments.js';
-
-const AMOUNT_HASH_TAGS = 5;
-const LENGTH_COMMENTS = 140;
+import { onDocumentKeyDown } from './utils.js';
+import { validationHashTags } from './validation-hash-tags.js';
+import { validationComments } from './validation-comments.js';
 
 const uploadPictureForm = document.querySelector('.img-upload__form');
 const editPicturePopup = uploadPictureForm.querySelector('.img-upload__overlay');
@@ -21,12 +18,7 @@ const closeEditPopup = () => {
   commentsInput.value = '';
 };
 
-const onEditPopupEscKeyDown = (evt) => {
-  if (isEscEvent(evt)) {
-    evt.preventDefault();
-    closeEditPopup();
-  }
-};
+const onEditPopupEscKeyDown = (evt) => onDocumentKeyDown(evt, closeEditPopup);
 
 const onCloseEditPictureButtonClick = () => closeEditPopup();
 
@@ -41,18 +33,10 @@ const openEditPopup = () => {
 
 uploadPictureInput.addEventListener('change', () => openEditPopup());
 
-//Валидация поля для заполнения Хештегов
-hashtagsInput.addEventListener('change', () => {
-  if (!matchingPattern()) {
-    hashtagsInput.setCustomValidity('Проверьте правильность написания хештегов: хештег должен начинаться с #, содержать только буквы и числа и не превышать 20 символов');
-  } else if (isDuplicate()) {
-    hashtagsInput.setCustomValidity('Один и тот же хэш-тег не может быть использован дважды');
-  } else if (isTooMushHashTags()) {
-    hashtagsInput.setCustomValidity(`Уменьшите количество хэш-тегов на ${createHashTags().length - AMOUNT_HASH_TAGS}`);
-  } else {
-    hashtagsInput.setCustomValidity('');
-  }
-});
+// Заполнение полей формы
+
+//Поле ввода хэш-тэгов
+hashtagsInput.addEventListener('change', () => validationHashTags());
 
 // Отмена закрытия формы при заполнении поля хеш-тегов
 hashtagsInput.addEventListener('focus', () => {
@@ -63,14 +47,8 @@ hashtagsInput.addEventListener('blur', () => {
   document.addEventListener('keydown', onEditPopupEscKeyDown);
 });
 
-//Валидация поля для заполнения комментариев
-commentsInput.addEventListener('input', () => {
-  if (!isValidityLengthComments()) {
-    commentsInput.setCustomValidity(`Уменьшите длину комментария на ${commentsInput.value.length - LENGTH_COMMENTS} симв.`);
-  } else {
-    commentsInput.setCustomValidity('');
-  }
-});
+//Поле ввода комментариев
+commentsInput.addEventListener('input', () => validationComments());
 
 // Отмена закрытия формы при заполнении поля комментариев
 commentsInput.addEventListener('focus', () => {
@@ -81,4 +59,4 @@ commentsInput.addEventListener('blur', () => {
   document.addEventListener('keydown', onEditPopupEscKeyDown);
 });
 
-export { AMOUNT_HASH_TAGS, LENGTH_COMMENTS, hashtagsInput, commentsInput };
+export { hashtagsInput, commentsInput };
