--- js/test-comments.js
@@ -1,10 +1,0 @@
-import { checkStringLength } from './utils.js';
-import { commentsInput, LENGTH_COMMENTS } from './form.js';
-
-// Проверка длины комментария
-const isValidityLengthComments = () => {
-  const comments = commentsInput.value;
-  return checkStringLength(comments, LENGTH_COMMENTS);
-};
-
-export { isValidityLengthComments };
